original_research/scripts/llm_analysis/requirements.txt
langchain-core>=0.3
langchain>=0.3
langchain-openai>=0.3
langchain-google-genai>=2.0
langchain-anthropic>=0.3
langchain-xai>=0.2
langchain-groq>=0.2
langchain-aws>=0.2
Pillow>=10.0
python-dotenv>=1.0
